refactor: separate integration tests
Also sets up CI to run tests.
