+++ crates/executor/host/src/lib.rs
@@ -189,75 +189,3 @@ impl<T: Transport + Clone, P: Provider<T> + Clone> HostExecutor<T, P> {
         Ok(client_input)
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    use alloy_provider::ReqwestProvider;
-    use rsp_client_executor::ClientExecutor;
-    use tracing_subscriber::{
-        filter::EnvFilter, fmt, prelude::__tracing_subscriber_SubscriberExt,
-        util::SubscriberInitExt,
-    };
-    use url::Url;
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn test_e2e_ethereum() {
-        run_e2e::<EthereumVariant>("RPC_1", 18884864, "client_input_ethereum.json").await;
-    }
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn test_e2e_optimism() {
-        run_e2e::<OptimismVariant>("RPC_10", 122853660, "client_input_optimism.json").await;
-    }
-
-    #[test]
-    fn test_input_bincode_roundtrip() {
-        let file = std::fs::File::open("client_input_ethereum.json").unwrap();
-        let client_input: ClientExecutorInput = serde_json::from_reader(file).unwrap();
-        let serialized = bincode::serialize(&client_input).unwrap();
-        let deserialized = bincode::deserialize::<ClientExecutorInput>(&serialized).unwrap();
-        assert_eq!(client_input, deserialized);
-    }
-
-    async fn run_e2e<V>(env_var_key: &str, block_number: u64, input_file: &str)
-    where
-        V: Variant,
-    {
-        // Intialize the environment variables.
-        dotenv::dotenv().ok();
-
-        // Initialize the logger.
-        let _ = tracing_subscriber::registry()
-            .with(fmt::layer())
-            .with(EnvFilter::from_default_env())
-            .try_init();
-
-        // Setup the provider.
-        let rpc_url =
-            Url::parse(std::env::var(env_var_key).unwrap().as_str()).expect("invalid rpc url");
-        let provider = ReqwestProvider::new_http(rpc_url);
-
-        // Setup the host executor.
-        let host_executor = HostExecutor::new(provider);
-
-        // Execute the host.
-        let (client_input, _) =
-            host_executor.execute(block_number).await.expect("failed to execute host");
-
-        // Setup the client executor.
-        let client_executor = ClientExecutor;
-
-        // Execute the client.
-        client_executor.execute::<V>(client_input.clone()).expect("failed to execute client");
-
-        // Save the client input to a file.
-        let file = std::fs::File::create(input_file).unwrap();
-        serde_json::to_writer_pretty(file, &client_input).unwrap();
-
-        // Load the client input from a file.
-        let file = std::fs::File::open(input_file).unwrap();
-        let _: ClientExecutorInput = serde_json::from_reader(file).unwrap();
-    }
-}
